experiments/neuro_symbolic_gate/requirements.txt
pyyaml>=6.0
openai>=1.0
pandas>=2.0
openpyxl>=3.0
clingo>=5.6          # ASP solver backend (rsafety_v0.lp)
